Add sponsors field to event model
Additionally, change the event import task to add the sponsors to events imported from Blueprint.

To better reflect the data as it was in Blueprint we made a decision to store existing sponsors as an associated_pages field on the event model. In Blueprint the relationship was between events and sponsors - going forward the relationship will be between event series and sponsors.