Add video file migration to Recording import task
The plan:
When adding a recording, find all the video files in the backup_data that relate.
Find the best quality of these files by checking their encoding formats.
Find that file in the Wheeler Centre S3 bucket.
Migrate this to Youtube using the client API - currently this doesn’t work because the client API can’t use a file location on a remote server; the file must be local.
Associate the Recording with the uploaded Youtube file.