Fix issue with cached search values
The problem was that the masthead caching is encompassing the search value, resulting in incorrect values displaying in the search field.

To fix this, create a hidden input outside of the masthead caching with the search query as its value. Update the masthead and footer search fields with this value using javascript.